requirements.txt
cassandra-driver==3.29.2
click==8.1.8
Faker==37.1.0
geomet==0.2.1.post1
kafka-python==2.1.5
py4j==0.10.9.7
pyspark==3.5.5
six==1.17.0
tzdata==2025.2
